add setupHook util module
For babel-plugin-css-modules-transform to use without
registering a node hook:

https://github.com/michalkvasnicak/babel-plugin-css-modules-transform/blob/master/src/index.js#L167
